refactor: localize unsigned data decoding
